Replace atomics with a mutex & do reloading of marian inside worker
Replaced two atomic pointers with a mutex and locked sections to remove potential race condition. Also moved changing the marian model to an operation the worker can perform so the main thread does not have to wait for it. This removes the need to disconnect/reconnect to a new MarianInterface. Sentences can be queued while the worker is loading another model but not before the worker has loaded any model.