Skip disk entirely when extracting from download
We already download to a QByteArray. Not really useful to write that first to disk, and then read it into memory again. If relying on the OS' memory paging for the downloaded file becomes an issue, we could rewrite the download to progressively write to disk and change extractTarGz back again.