Don't launch a new console on windows
Windows applications built through cmake are not GUI by default. Doing this will make sure that we don't launch a console. As a downside when we add command line interface, it'd probably require different approach to make it work on windows, but then again most people on windows don't use command line...