Check for native messaging call pattern
Previously we just looked for known extension ids, but in case that wasn't found we'd launch with GUI and everything. If a manifest file would go out of sync and allow for translateLocally to be called from the browser, but translateLocally didn't know of the extension, it would just pop up. Pretty unexpected (but it did have the nice side effect of fixing the manifest files for known browsers, so that was something at least...)