Remove RepoManager
I'm still keeping a copy of the repositories in ModelManager in a QMap for quick key lookups since those are used for name lookups in the repo table and the model dropdown and accessed "frequently". Should they have been a member variable of `Model` instead? Maybe, but we don't keep a tree of unique models and repository pointers in memory that we try to keep in sync, we just have a simple flat structure coming out of Settings.