Do not make the modelMeta.json hidden
The model files are already extracted in a semi-hidden config directory that only an advanced user would check. It is only a bit confusing if we have an extra hidden file in there (by prefixing it with a dot) when debugging. Furthermore the dot prefix does nothing on Windows and it just makes it look weird.