Handle local and remote models from the languages tab (#113)
* WiP. Functional but several small bugs

* Merge ifelse

* Do not make the modelMeta.json hidden

The model files are already extracted in a semi-hidden config directory that only an advanced user would check. It is only a bit confusing if we have an extra hidden file in there (by prefixing it with a dot) when debugging. Furthermore the dot prefix does nothing on Windows and it just makes it look weird.

* Fix #114 . Also report builtin repo name from model_info.json

Beforehand we were not reading the ModelMeta after writing it, so any newly downloaded model would not have had that extra information available, resulting in id() not matching

* Make Model table sortable

* Use modelProxy_ when retrieving model by row index

* Add filtering table view

Pressing ctrl + f (or whatever the native key combo is on your machine) pops up a search field in the table view

* Merge download & update buttons

* Remove installed column, hide `-1` version numbers

* Hide Remote Version column by default

Otherwise it is just empty and non-descript

* Rename columns to `Version` and `Available` to be less wide

Also stretch (and more importantly, shrink!) the `type` column as that one is probably the one that varies most and is also the least important to be displaying its full content all the time.

* Revert the currentIndex thingy

* Make `ModelManager::data()` less repetitive

* Add a bit of *colour science*

Joking. But this at least works for light & dark mode.

Co-authored-by: Jelmer van der Linde <jelmer@ikhoefgeen.nl>